Add information about Reminder Project
Reminder is to cause to remember so the aim of this project is to develop a reminder application which can store the words  in the search engine to destroy the forgetfulness problem.Like a password or something you search.We are going to do search engine which have a word pool. So when people search anything the page , the search engine start with a reminder menu. For example , if I write a "Ho" , the search engine show up all words starting with H and it will display the words with the dropdown menu . Also , the search engine will save each word that you're looking for before.So you will not have any problems like what I've been looking for.